feat: create initial database schema including tables for users, roles, profiles, job ads, applications, CVs, and skills.

+++ database/migrations/2024_01_01_000000_create_base_schema_table.php
@@ -589,7 +589,7 @@ return new class extends Migration
         if (!Schema::hasTable('companies')) {
             Schema::create('companies', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedBigInteger('user_id')->unique();
+                $table->integer('user_id')->unique();
                 $table->string('name');
                 $table->string('slug')->unique();
                 $table->text('description')->nullable();
